on line

+++ blog/src/main/java/cn/edu/jxnu/blog/controller/admin/CommentAdminController.java
@@ -1,14 +1,13 @@
 package cn.edu.jxnu.blog.controller.admin;
 
 import javax.annotation.Resource;
-import javax.servlet.http.HttpServletResponse;
 
-import org.springframework.stereotype.Controller;
+import org.slf4j.Logger;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import cn.edu.jxnu.blog.commons.ResponseUtil;
 import cn.edu.jxnu.blog.domin.Blog;
 import cn.edu.jxnu.blog.domin.Comment;
 import cn.edu.jxnu.blog.domin.PageBean;
@@ -23,10 +22,13 @@ import com.alibaba.fastjson.serializer.SerializerFeature;
 /**
  * @Description 博客访问控制层
  */
-@Controller
+@RestController
 @RequestMapping(value = "admin/comment")
 public class CommentAdminController {
 
+
+	private static final Logger log = org.slf4j.LoggerFactory
+			.getLogger(CommentAdminController.class);
 	@Resource
 	private BlogService blogService;
 	@Resource
@@ -37,8 +39,9 @@ public class CommentAdminController {
 	public String listByPage(
 			@RequestParam(value = "page", required = false) String page,
 			@RequestParam(value = "limit", required = false) String limit,
-			@RequestParam(value = "state", required = false) String state,
-			HttpServletResponse response) throws Exception {
+			@RequestParam(value = "state", required = false) String state)
+			throws Exception {
+		log.info("当前请求查看评论管理页面。。。");
 		PageBean<Comment> pageBean = new PageBean<Comment>(
 				Integer.parseInt(page), Integer.parseInt(limit));
 		pageBean.getMap().put("state", state);
@@ -53,23 +56,41 @@ public class CommentAdminController {
 		result.put("count", pageBean.getTotal());
 		result.put("data", jsonArray);
 		result.put("code", 0);// 封装接口，成功返回0
-		ResponseUtil.write(response, result);
-		return null;
+		return JSON.toJSONString(result);
 	}
 
 	// 删除评论
 	@RequestMapping(value = "/delete", method = RequestMethod.POST)
 	public String deleteComment(
-			@RequestParam(value = "ids", required = false) String ids,
-			HttpServletResponse response) throws Exception {
-		String[] idStr = ids.split(",");
-		for (int i = 0; i < idStr.length; i++) {
-			commentService.deleteComment(Integer.parseInt(idStr[i]));
-		}
+			@RequestParam(value = "ids", required = false) String ids)
+			throws Exception {
+		log.info("当前请求删除评论。。。");
 		JSONObject result = new JSONObject();
-		result.put("success", true);
-		ResponseUtil.write(response, result);
-		return null;
+		String[] idStr = ids.split(",");
+		int delete =0;
+		for (int i = 0; i < idStr.length; i++) {
+			if (commentService.getById(Integer.parseInt(idStr[i])) != null) {
+				Comment comment = commentService.getById(Integer.parseInt(idStr[i]));
+				if (comment.getState().equals(1)) { //删除的是审核 通过的
+					// 删除评论的时候，如果是删除已经审核的评论，则需要减小文章的评论数
+					Blog blog = comment.getBlog();
+					if(blog != null) { //审核通过且博客存在，需要减
+						delete = commentService.deleteComment(Integer.parseInt(idStr[i]));
+						blog.setReplyHit(blog.getReplyHit() - 1);
+						blogService.updateBlog(blog);
+					}else { //审核通过但是博客不存在了。直接删除
+						delete = commentService.deleteComment(Integer.parseInt(idStr[i]));
+					}
+				} else {
+					delete = commentService.deleteComment(Integer.parseInt(idStr[i]));
+				}
+
+			}
+		}
+		if (delete>0) {
+			result.put("success", true); // 删除成功
+		}
+		return JSON.toJSONString(result);
 	}
 
 	/**
@@ -81,11 +102,12 @@ public class CommentAdminController {
 	 * @return
 	 * @throws Exception
 	 */
-	@RequestMapping(value = "review",method=RequestMethod.POST)
+	@RequestMapping(value = "review", method = RequestMethod.POST)
 	public String reviewComment(
 			@RequestParam(value = "ids", required = false) String ids,
-			@RequestParam(value = "state", required = false) String state,
-			HttpServletResponse response) throws Exception {
+			@RequestParam(value = "state", required = false) String state)
+			throws Exception {
+		log.info("当前请求审核评论。。。");
 		String[] idStr = ids.split(",");
 		for (int i = 0; i < idStr.length; i++) {
 			Comment comment = new Comment();
@@ -94,13 +116,14 @@ public class CommentAdminController {
 			commentService.updateComment(comment); // 根据id更新评论的状态
 			Comment commentTemp = commentService.getById(Integer.parseInt(idStr[i]));
 			Blog blog = commentTemp.getBlog();// 根据id获取文章
-			blog.setReplyHit(blog.getReplyHit() + 1);
-			System.out.println("blog reply:"+blog.getReplyHit());
+			if (Integer.parseInt(state)==1) {
+				blog.setReplyHit(blog.getReplyHit() + 1);
+			}
+			System.out.println("blog reply:" + blog.getReplyHit());
 			blogService.updateBlog(blog);
 		}
 		JSONObject result = new JSONObject();
 		result.put("success", true);
-		ResponseUtil.write(response, result);
-		return null;
+		return JSON.toJSONString(result);
 	}
 }
